Rename of the live-clock component
